+++ ibmmq/cmqc_darwin.go
@@ -14,7 +14,7 @@ package ibmmq
 * for each value; those can be found in other header files such as
 * cmqc.h.
 ****************************************************************
-* Copyright (c) IBM Corporation 1993, 2022
+* Copyright (c) IBM Corporation 1993, 2023
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -30,8 +30,8 @@ package ibmmq
 ****************************************************************
 *
 *   <BEGIN_BUILDINFO>
-*   Generated on:  9/29/22 4:30 PM
-*   Build Level:   p931-L220929.1
+*   Generated on:  2/7/23 5:13 PM
+*   Build Level:   p932-L230207
 *   Build Type:    Production
 *   <END_BUILDINFO>
  */
@@ -173,6 +173,8 @@ const (
 	MQAUTH_SET_IDENTITY_CONTEXT    int32  = 16
 	MQAUTH_SUBSCRIBE               int32  = 20
 	MQAUTH_SYSTEM                  int32  = 22
+	MQAUTOCLUS_TYPE_NONE           int32  = 0
+	MQAUTOCLUS_TYPE_UNIFORM        int32  = 1
 	MQAUTO_START_NO                int32  = 0
 	MQAUTO_START_YES               int32  = 1
 	MQAXC_CURRENT_LENGTH           int32  = 424
@@ -304,8 +306,6 @@ const (
 	MQCACF_CF_STRUC_NAMES          int32  = 3095
 	MQCACF_CHAR_ATTRS              int32  = 3189
 	MQCACF_CHILD_Q_MGR_NAME        int32  = 3040
-	MQCACF_CLUS_CHAN_Q_MGR_NAME    int32  = 5507
-	MQCACF_CLUS_SHORT_CONN_NAME    int32  = 5508
 	MQCACF_COMMAND_MQSC            int32  = 3075
 	MQCACF_COMMAND_SCOPE           int32  = 3080
 	MQCACF_CONFIGURATION_DATE      int32  = 3091
@@ -354,14 +354,21 @@ const (
 	MQCACF_LAST_PUB_TIME           int32  = 3162
 	MQCACF_LAST_PUT_DATE           int32  = 3128
 	MQCACF_LAST_PUT_TIME           int32  = 3129
-	MQCACF_LAST_USED               int32  = 3211
+	MQCACF_LAST_USED               int32  = 3220
 	MQCACF_LOCAL_Q_NAMES           int32  = 3015
 	MQCACF_LOG_PATH                int32  = 3074
+	MQCACF_LOG_START_DATE          int32  = 3214
+	MQCACF_LOG_START_LSN           int32  = 3215
+	MQCACF_LOG_START_TIME          int32  = 3216
 	MQCACF_MEDIA_LOG_EXTENT_NAME   int32  = 3073
 	MQCACF_MODEL_Q_NAMES           int32  = 3016
 	MQCACF_MQCB_NAME               int32  = 3192
 	MQCACF_NAMELIST_NAMES          int32  = 3013
+	MQCACF_NHA_GROUP_INITIAL_DATE  int32  = 3217
+	MQCACF_NHA_GROUP_INITIAL_LSN   int32  = 3218
+	MQCACF_NHA_GROUP_INITIAL_TIME  int32  = 3219
 	MQCACF_NHA_INSTANCE_NAME       int32  = 3211
+	MQCACF_NHA_REPL_ADDRESS        int32  = 3220
 	MQCACF_NONE                    int32  = 3171
 	MQCACF_OBJECT_NAME             int32  = 3046
 	MQCACF_OBJECT_Q_MGR_NAME       int32  = 3023
@@ -378,6 +385,7 @@ const (
 	MQCACF_PUT_DATE                int32  = 3137
 	MQCACF_PUT_TIME                int32  = 3138
 	MQCACF_Q_MGR_CPF               int32  = 3076
+	MQCACF_Q_MGR_DATA_PATH         int32  = 3212
 	MQCACF_Q_MGR_START_DATE        int32  = 3175
 	MQCACF_Q_MGR_START_TIME        int32  = 3176
 	MQCACF_Q_MGR_UOW_ID            int32  = 3086
@@ -461,6 +469,7 @@ const (
 	MQCACF_TO_SUB_NAME             int32  = 3164
 	MQCACF_TO_TOPIC_NAME           int32  = 3149
 	MQCACF_TRANSACTION_ID          int32  = 3085
+	MQCACF_UNIFORM_CLUSTER_NAME    int32  = 3213
 	MQCACF_UOW_LOG_EXTENT_NAME     int32  = 3064
 	MQCACF_UOW_LOG_START_DATE      int32  = 3062
 	MQCACF_UOW_LOG_START_TIME      int32  = 3063
@@ -1023,7 +1032,7 @@ const (
 	MQCMDI_SEC_SIGNOFF_ERROR       int32  = 17
 	MQCMDI_SEC_TIMER_ZERO          int32  = 14
 	MQCMDI_SEC_UPPERCASE           int32  = 21
-	MQCMDL_CURRENT_LEVEL           int32  = 931
+	MQCMDL_CURRENT_LEVEL           int32  = 932
 	MQCMDL_LEVEL_1                 int32  = 100
 	MQCMDL_LEVEL_101               int32  = 101
 	MQCMDL_LEVEL_110               int32  = 110
@@ -1072,6 +1081,7 @@ const (
 	MQCMDL_LEVEL_925               int32  = 925
 	MQCMDL_LEVEL_930               int32  = 930
 	MQCMDL_LEVEL_931               int32  = 931
+	MQCMDL_LEVEL_932               int32  = 932
 	MQCMD_ACCOUNTING_MQI           int32  = 167
 	MQCMD_ACCOUNTING_Q             int32  = 168
 	MQCMD_ACTIVITY_MSG             int32  = 69
@@ -1622,6 +1632,10 @@ const (
 	MQFIELD_WQR_StrucId            int32  = 8000
 	MQFIELD_WQR_StrucLength        int32  = 8002
 	MQFIELD_WQR_Version            int32  = 8001
+	MQFSENC_NO                     int32  = 0
+	MQFSENC_UNKNOWN                int32  = 2
+	MQFSENC_YES                    int32  = 1
+	MQFS_SHARED                    int32  = -1
 	MQFUN_TYPE_COMMAND             int32  = 5
 	MQFUN_TYPE_JVM                 int32  = 1
 	MQFUN_TYPE_PROCEDURE           int32  = 3
@@ -1740,6 +1754,7 @@ const (
 	MQIACF_AUTH_REC_TYPE           int32  = 1412
 	MQIACF_AUTH_REMOVE_AUTHS       int32  = 1117
 	MQIACF_AUTH_SERVICE_ATTRS      int32  = 1264
+	MQIACF_AUTO_CLUSTER_TYPE       int32  = 1448
 	MQIACF_AUX_ERROR_DATA_INT_1    int32  = 1070
 	MQIACF_AUX_ERROR_DATA_INT_2    int32  = 1071
 	MQIACF_BACKOUT_COUNT           int32  = 1241
@@ -1801,6 +1816,8 @@ const (
 	MQIACF_CTL_OPERATION           int32  = 1366
 	MQIACF_CUR_MAX_FILE_SIZE       int32  = 1438
 	MQIACF_CUR_Q_FILE_SIZE         int32  = 1437
+	MQIACF_DATA_FS_IN_USE          int32  = 1449
+	MQIACF_DATA_FS_SIZE            int32  = 1450
 	MQIACF_DB2_CONN_STATUS         int32  = 1150
 	MQIACF_DELETE_OPTIONS          int32  = 1092
 	MQIACF_DESTINATION_CLASS       int32  = 1273
@@ -1839,13 +1856,19 @@ const (
 	MQIACF_INVALID_DEST_COUNT      int32  = 1371
 	MQIACF_ITEM_COUNT              int32  = 1378
 	MQIACF_KNOWN_DEST_COUNT        int32  = 1369
-	MQIACF_LAST_USED               int32  = 1444
+	MQIACF_LAST_USED               int32  = 1465
 	MQIACF_LDAP_CONNECTION_STATUS  int32  = 1409
 	MQIACF_LISTENER_ATTRS          int32  = 1222
 	MQIACF_LISTENER_STATUS_ATTRS   int32  = 1223
 	MQIACF_LOG_COMPRESSION         int32  = 1322
+	MQIACF_LOG_EXTENT_SIZE         int32  = 1451
+	MQIACF_LOG_FS_IN_USE           int32  = 1452
+	MQIACF_LOG_FS_SIZE             int32  = 1453
 	MQIACF_LOG_IN_USE              int32  = 1420
+	MQIACF_LOG_PRIMARIES           int32  = 1454
 	MQIACF_LOG_REDUCTION           int32  = 1422
+	MQIACF_LOG_SECONDARIES         int32  = 1455
+	MQIACF_LOG_TYPE                int32  = 1456
 	MQIACF_LOG_UTILIZATION         int32  = 1421
 	MQIACF_MAX_ACTIVITIES          int32  = 1236
 	MQIACF_MCAST_REL_INDICATOR     int32  = 1351
@@ -1867,6 +1890,12 @@ const (
 	MQIACF_MSG_TYPE                int32  = 1249
 	MQIACF_MULC_CAPTURE            int32  = 1324
 	MQIACF_NAMELIST_ATTRS          int32  = 1004
+	MQIACF_NHA_INSTANCE_ACTV_CONNS int32  = 1457
+	MQIACF_NHA_INSTANCE_BACKLOG    int32  = 1458
+	MQIACF_NHA_INSTANCE_IN_SYNC    int32  = 1459
+	MQIACF_NHA_INSTANCE_ROLE       int32  = 1460
+	MQIACF_NHA_IN_SYNC_INSTANCES   int32  = 1461
+	MQIACF_NHA_TOTAL_INSTANCES     int32  = 1462
 	MQIACF_NUM_PUBS                int32  = 1396
 	MQIACF_OBJECT_TYPE             int32  = 1016
 	MQIACF_OBSOLETE_MSGS           int32  = 1310
@@ -1912,10 +1941,16 @@ const (
 	MQIACF_Q_MGR_DQM               int32  = 1124
 	MQIACF_Q_MGR_EVENT             int32  = 1123
 	MQIACF_Q_MGR_FACILITY          int32  = 1231
+	MQIACF_Q_MGR_FS_ENCRYPTED      int32  = 1463
+	MQIACF_Q_MGR_FS_IN_USE         int32  = 1464
+	MQIACF_Q_MGR_FS_SIZE           int32  = 1465
 	MQIACF_Q_MGR_NUMBER            int32  = 1148
 	MQIACF_Q_MGR_PUBSUB            int32  = 1291
 	MQIACF_Q_MGR_STATUS            int32  = 1149
 	MQIACF_Q_MGR_STATUS_ATTRS      int32  = 1229
+	MQIACF_Q_MGR_STATUS_INFO_NHA   int32  = 1447
+	MQIACF_Q_MGR_STATUS_INFO_Q_MGR int32  = 1446
+	MQIACF_Q_MGR_STATUS_INFO_TYPE  int32  = 1445
 	MQIACF_Q_MGR_STATUS_LOG        int32  = 1415
 	MQIACF_Q_MGR_SYSTEM            int32  = 1122
 	MQIACF_Q_MGR_TYPE              int32  = 1085
@@ -2727,6 +2762,9 @@ const (
 	MQLDAP_AUTHORMD_SRCHGRPSN      int32  = 3
 	MQLDAP_NESTGRP_NO              int32  = 0
 	MQLDAP_NESTGRP_YES             int32  = 1
+	MQLOGTYPE_CIRCULAR             int32  = 0
+	MQLOGTYPE_LINEAR               int32  = 1
+	MQLOGTYPE_REPLICATED           int32  = 2
 	MQLR_AUTO                      int32  = -1
 	MQLR_MAX                       int32  = -2
 	MQLR_ONE                       int32  = 1
@@ -2871,6 +2909,14 @@ const (
 	MQMULC_REFINED                 int32  = 1
 	MQMULC_STANDARD                int32  = 0
 	MQNC_MAX_NAMELIST_NAME_COUNT   int32  = 256
+	MQNHABACKLOG_UNKNOWN           int32  = -1
+	MQNHACONNACTV_NO               int32  = 0
+	MQNHACONNACTV_YES              int32  = 1
+	MQNHAINSYNC_NO                 int32  = 0
+	MQNHAINSYNC_YES                int32  = 1
+	MQNHAROLE_ACTIVE               int32  = 1
+	MQNHAROLE_REPLICA              int32  = 2
+	MQNHAROLE_UNKNOWN              int32  = 0
 	MQNPMS_FAST                    int32  = 2
 	MQNPMS_NORMAL                  int32  = 1
 	MQNPM_CLASS_HIGH               int32  = 10
@@ -3414,6 +3460,7 @@ const (
 	MQRCCF_NETBIOS_NAME_ERROR      int32  = 3093
 	MQRCCF_NET_PRIORITY_ERROR      int32  = 4088
 	MQRCCF_NET_PRIORITY_WRONG_TYPE int32  = 4089
+	MQRCCF_NHA_NOT_AVAILABLE       int32  = 4098
 	MQRCCF_NONE_FOUND              int32  = 3200
 	MQRCCF_NOT_AUTHORIZED          int32  = 3081
 	MQRCCF_NOT_REGISTERED          int32  = 3073
@@ -3474,6 +3521,7 @@ const (
 	MQRCCF_Q_MGR_CCSID_ERROR       int32  = 3086
 	MQRCCF_Q_MGR_NAME_ERROR        int32  = 3074
 	MQRCCF_Q_MGR_NOT_IN_QSG        int32  = 3212
+	MQRCCF_Q_MGR_STATUS_NOT_FOUND  int32  = 4099
 	MQRCCF_Q_NAME_ERROR            int32  = 3076
 	MQRCCF_Q_STATUS_NOT_FOUND      int32  = 4091
 	MQRCCF_Q_TYPE_ERROR            int32  = 3022
@@ -4962,6 +5010,7 @@ const (
 	MQ_FUNCTION_LENGTH             int32  = 4
 	MQ_GROUP_ADDRESS_LENGTH        int32  = 264
 	MQ_GROUP_ID_LENGTH             int32  = 24
+	MQ_HOST_NAME_LENGTH            int32  = 256
 	MQ_INITIAL_KEY_LENGTH          int32  = 256
 	MQ_INSTALLATION_DESC_LENGTH    int32  = 64
 	MQ_INSTALLATION_NAME_LENGTH    int32  = 16
@@ -4980,6 +5029,7 @@ const (
 	MQ_LOG_EXTENT_NAME_LENGTH      int32  = 24
 	MQ_LOG_PATH_LENGTH             int32  = 1024
 	MQ_LRSN_LENGTH                 int32  = 12
+	MQ_LSN_LENGTH                  int32  = 64
 	MQ_LTERM_OVERRIDE_LENGTH       int32  = 8
 	MQ_LUWID_LENGTH                int32  = 16
 	MQ_LU_NAME_LENGTH              int32  = 8
@@ -5001,6 +5051,7 @@ const (
 	MQ_NAMELIST_DESC_LENGTH        int32  = 64
 	MQ_NAMELIST_NAME_LENGTH        int32  = 48
 	MQ_NHA_INSTANCE_NAME_LENGTH    int32  = 48
+	MQ_NHA_REPL_ADDRESS_LENGTH     int32  = 264
 	MQ_OBJECT_INSTANCE_ID_LENGTH   int32  = 24
 	MQ_OBJECT_NAME_LENGTH          int32  = 48
 	MQ_OPERATOR_MESSAGE_LENGTH     int32  = 4
@@ -5021,6 +5072,7 @@ const (
 	MQ_QSG_NAME_LENGTH             int32  = 4
 	MQ_Q_DESC_LENGTH               int32  = 64
 	MQ_Q_MGR_CPF_LENGTH            int32  = 4
+	MQ_Q_MGR_DATA_PATH_LENGTH      int32  = 1024
 	MQ_Q_MGR_DESC_LENGTH           int32  = 64
 	MQ_Q_MGR_IDENTIFIER_LENGTH     int32  = 48
 	MQ_Q_MGR_NAME_LENGTH           int32  = 48